requirements.txt
requests
pandas
numpy
json
pathlib
beautifulsoup4
psycopg2
sqlalchemy
apache-airflow
apache-airflow-providers-ssh